translate_env/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
python-decouple==3.3
requests==2.24.0
urllib3==1.25.10
